+++ Tree_Controller/Tools/TreeControllerTool.cs
@@ -9,6 +9,7 @@ namespace Tree_Controller.Tools
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Runtime.CompilerServices;
     using Colossal.Annotations;
     using Colossal.Entities;
     using Colossal.Logging;
@@ -518,6 +519,7 @@ namespace Tree_Controller.Tools
                         m_TreeDataLookup = SystemAPI.GetComponentLookup<TreeData>(isReadOnly: true),
                         m_TreeLookup = SystemAPI.GetComponentLookup<Tree>(isReadOnly: true),
                         m_VegetationLookup = SystemAPI.GetComponentLookup<Vegetation>(isReadOnly: true),
+                        m_SubMeshLookup = SystemAPI.GetBufferLookup<SubMesh>(isReadOnly: true),
                     };
                     inputDeps = JobChunkExtensions.ScheduleParallel(changeTreeAgeWithinRadiusJob, m_VegetationQuery, inputDeps);
                     m_ToolOutputBarrier.AddJobHandleForProducer(inputDeps);
@@ -773,6 +775,7 @@ namespace Tree_Controller.Tools
             public ComponentLookup<Tree> m_TreeLookup;
             public ComponentLookup<TreeData> m_TreeDataLookup;
             public ComponentLookup<Vegetation> m_VegetationLookup;
+            public BufferLookup<Game.Prefabs.SubMesh> m_SubMeshLookup;
 
             /// <summary>
             /// Executes job which will change state or prefab for trees within a radius.
@@ -792,13 +795,21 @@ namespace Tree_Controller.Tools
                     if (CheckForHoveringOverTree(m_Position, transformNativeArray[i].m_Position, m_Radius))
                     {
                         Entity currentEntity = entityNativeArray[i];
-                        if (m_OverrideState && m_TreeLookup.HasComponent(currentEntity) && m_OverridePrefab == false)
+                        if (m_OverrideState &&
+                            m_TreeLookup.HasComponent(currentEntity) &&
+                            m_OverridePrefab == false)
                         {
                             Game.Objects.Tree currentTreeData = treeNativeArray[i];
-                            currentTreeData.m_State = GetTreeState(m_Ages, currentTreeData);
-                            buffer.SetComponent(unfilteredChunkIndex, currentEntity, currentTreeData);
-                            buffer.AddComponent<RecentlyChanged>(unfilteredChunkIndex, currentEntity);
-                            buffer.AddComponent<BatchesUpdated>(unfilteredChunkIndex, currentEntity);
+                            TreeState newState = GetTreeState(m_Ages, currentTreeData);
+                            if (newState != TreeState.Stump ||
+                               (m_SubMeshLookup.TryGetBuffer(prefabRefNativeArray[i].m_Prefab, out DynamicBuffer<SubMesh> subMeshBuffer) &&
+                                subMeshBuffer.Length > 5))
+                            {
+                                currentTreeData.m_State = newState;
+                                buffer.SetComponent(unfilteredChunkIndex, currentEntity, currentTreeData);
+                                buffer.AddComponent<RecentlyChanged>(unfilteredChunkIndex, currentEntity);
+                                buffer.AddComponent<BatchesUpdated>(unfilteredChunkIndex, currentEntity);
+                            }
                             continue;
                         }
 
@@ -835,12 +846,20 @@ namespace Tree_Controller.Tools
                                 buffer.RemoveComponent<Tree>(unfilteredChunkIndex, currentEntity);
                             }
 
-                            // Override state of existing tree and prefab.
-                            else if (m_OverrideState && m_TreeDataLookup.HasComponent(currentPrefabRef.m_Prefab) && m_TreeLookup.HasComponent(currentEntity))
+                            // Override state of existing tree
+                            else if (m_OverrideState &&
+                                     m_TreeDataLookup.HasComponent(currentPrefabRef.m_Prefab) &&
+                                     m_TreeLookup.HasComponent(currentEntity))
                             {
                                 Game.Objects.Tree currentTreeData = treeNativeArray[i];
-                                currentTreeData.m_State = GetTreeState(m_Ages, currentTreeData);
-                                buffer.SetComponent(unfilteredChunkIndex, currentEntity, currentTreeData);
+                                TreeState newState = GetTreeState(m_Ages, currentTreeData);
+                                if (newState != TreeState.Stump ||
+                                   (m_SubMeshLookup.TryGetBuffer(currentPrefabRef.m_Prefab, out DynamicBuffer<SubMesh> subMeshBuffer)  &&
+                                    subMeshBuffer.Length > 5))
+                                {
+                                    currentTreeData.m_State = newState;
+                                    buffer.SetComponent(unfilteredChunkIndex, currentEntity, currentTreeData);
+                                }
                             }
 
                             buffer.SetComponent(unfilteredChunkIndex, currentEntity, currentPrefabRef);