+++ Tree_Controller/Properties/Release/ChangeLog.md
@@ -1,2 +1,5 @@
-﻿# Patch v1.6.13
-* Fix Incorrect Season IDs resulting in log spam and occasional error pop-up. Also fixes whatever functionality depended on that method.
+﻿# Patch v1.6.13.100
+* Radius for Tree Changing Tools can go down to 1 now.
+* Some aspects of the mod have been expanded from prefabs in the Vegetation tab to all prefabs with PlantData.
+* Improved handling of prefab entities being regenerated from scratch for compatiblity with other mods.
+* Added stump to Tree Age Changer tool when option is selected.